pyproject.toml
[project]
name = "ai-costing-graph"
version = "0.1.0"
description = "RAG pipeline for generating legal costing documents using Ollama, Langchain, and Neo4j"
authors = [
    {name = "Brendan Coen"}
]
dependencies = [
    "langchain>=0.1.0",
    "langchain-community>=0.0.10",
    "langchain-experimental>=0.0.10",
    "neo4j>=5.14.0",
    "ollama>=0.1.0",
    "langchain-neo4j>=0.0.1",
    "sentence-transformers>=2.2.2",
    "pypdf>=3.17.0",
    "python-docx>=1.0.0",
    "unstructured>=0.10.30",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "rich>=13.7.0",
    "typer>=0.9.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "B", "I"]
ignore = []

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=src" 